Further refine libo_CHECK_SYSTEM_MODULE
Adds the possibility to have default-disabled externals and
various fixed variants. The reorganized option handling is also
easier to follow. Can't use the newer m4_case, so this uses m4_if
with multiple compares, but the code difference is minimal.

This also swaps the 4th and 5th argument to reflect the precedence
of "enabled" over the "system" handling. And since these now have
multiple values, TRUE was replaces with more sensible strings.
These and the other arguments are now checked, as possible.

I've also thought about dropping the 2nd argument for m4_toupper,
but that changes to much for just a minimal gain.

Change-Id: I22e835d81f9288f22d42be36e939374f7a455599
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/127424
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>