Resolves: tdf#146377 Propagate condition error of IF(condition)
Previsouly a given ELSE-path was executed as the condition is never
TRUE on error, which wasn't significant unless the ELSE was an
error evaluating function like IFERROR() or ISERROR() or other
IS...() functions.

This changes behaviour of an undocumented side effect, so *might*
break existing expressions that (unwillingly?) relied on it.

Gnumeric, Excel and GoogleSheets all agree on propagating the
error value as implemented now.

Change-Id: I0905a363be7a3b02925f1c41b4e6c8d36f12df93
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/127515
Reviewed-by: Eike Rathke <erack@redhat.com>
Tested-by: Jenkins