tdf#146454 - Check for multiline string in setDataArray
Check for a multiline string in the setDataArray function. If that is
the case, set a string or edit cell. Otherwise, just set a plain string in order to prevent performance issues due to the adjustments of the row height.

Regression from b9e2c5d2a36fbd189c20448cadf4212edf02914d

Change-Id: I9216b198b878fda5da29596d25504db853acd756
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/127660
Tested-by: Jenkins
Reviewed-by: Andreas Heinisch <andreas.heinisch@yahoo.de>