Outline folding: Check frame area is valid before using pointer to
saved outline frame

It seems when a frame is removed it is not immediately destroyed, so a
pointer to the removed frame may not be a null pointer. This patch
checks for valid frame area definition before using the save outline
frame pointer with intention to fix an intermittent crash occurrence
when an outline paragraph is removed.

Change-Id: I4f9b55acc6aeb83553a3f26c2701159c2ba439a2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/127717
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>