tdf#146084 No hyphenation warning for Vietnamese
The Vietnamese language does not use hyphenation, so LO should not
warn for lack of hyphenation data. This was described in the commit
message of 151c56ed547490a99d912524c0e56b5d6d4a1939, but the actual
code for fixing the problem for Vietnamese langauge was missing.
This patch, adds the Vietnamese language to the list of languages
that do not use hyphenation, so the warning for the lack of
hyphenation data will no longer be displayed for this language.

As described in the above commit:

"In Vietnamese all words consist of single syllables, so they are
often very short; hyphenation is not allowed at all."
https://tug.org/TUGboat/tb29-1/tb91thanh-vntex.pdf
Hyphenation is declined in Vietnamese orthography since 1975
https://www.quora.com/When-did-hyphenation-decline-in-Vietnamese-orthography

Change-Id: Id8d3681ad0a93065e61e9da2422af64793fbc111
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/128177
Tested-by: Jenkins
Reviewed-by: Adolfo Jayme Barrientos <fitojb@ubuntu.com>