SwNavigator: fix outline delete removes first character of next outline
paragraph

The to be deleted outline selection may already be to the start of the
following outline paragraph as happens when a table is the last
content of the to be deleted outline. In this case do not extend the
selection right or the first character of the following outline
paragraph will be removed. Also check if no selection was made which
indicates an empty paragraph and the selection right is needed.

Change-Id: I5b2516bd7b1e88ee415986dbe9313f93039dbac0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/128264
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>