split gbuild's <module>.all to <module>.allbuild or .allcheck
This is similar to the removal of the 'all' target in the past.
Using 'make vcl.all' is supposed to build the vcl module and all
its dependencies, the problem is that it builds also tests, which
means not only building and running the tests but also their
dependencies too, so 'make vcl.all' may in fact also build Writer
because a unittest depends on something there.

So now <module>.allcheck is the new name for that, and newly
there's also the (IMO actually useful) <module>.allbuild .

Change-Id: I55baa014a657783e641cee67948ee8fb062b1982
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/128349
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>