tdf#146367 reintroduce approxAdd() similar handling with last summand
... to tie result to 0 eliminating opposite sign small equalish
fractions.

This does not "fix" all interim values of the bug's sample document,
which repeatedly adds single values in sequence using SUM() and should
use the + operator instead to benefit from approxAdd() for each, but
maintains the better accuracy of Kahan summation in the not near equal
zero cases. And the dreaded SUM(0.1;0.2;-0.3) also works again resulting
in 0 with all permutations.

Change-Id: I67188a6abbdb98212f070166ad02319c3d477e28
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/128451
Reviewed-by: Eike Rathke <erack@redhat.com>
Tested-by: Jenkins