UI Workaround for tdf#146653: change localized Source Han {Sans,Serif} CN...
...to Source Han Sans CN Normal and Source Han Serif CN Medium for Simplified Chinese UI, so that
the font shown is not bold.

Tranditional Chinese and Japanese settings are not touched. It is up to the
zh-TW and JA l10n team to decide whether to apply the same change.

This does not resolve the bug itself, but makes the UI usable when the
system has Source Han Sans CN or Source Han Serif CN fonts installed,
so that the default font used is normal weight rather than bold.

Change-Id: I52753448d886854d92ba8a26d8b03bf42522735f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/128798
Tested-by: Jenkins
Reviewed-by: Franklin Weng <franklin@goodhorse.idv.tw>
Reviewed-by: Eike Rathke <erack@redhat.com>