svx textcharacterspacingcontrol: allow smaller kerning
On of the options in the kerning list is "very tight"
which sets a -3 pt kerning. However, the custom spin
button only allows down to -2 pt. Certainly we need to allow
the user to adjust as much as the other options allow,
and it only makes sense to allow as much negative
kerning as positive kerning.
Additionally, this field is supposed to show the
REAL current value - which it can't do if the range
of values is too limited.
Finally, we can't pre-focus "very tight" very easily
if that spin button cannot match the "very tight" value.

[Based on testing, a too-small kerning gets treated
as the smallest-possible-kerning and doesn't cause
real problems.]

In terms of positive kerning, the real maximum pt value seems
to be 1638 before the denormalized version of it hits the sal_Int16
maximum and then starts wrapping over into minimum values.

Change-Id: I465fcd0bbf3ca310b203004b46632a9a1ee01a6f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/128984
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>