sw uno:Color: directly goto CHAR_DLG fonteffects tab
.uno:Color started the character properties dialog,
but opened whatever tab was last opened,
and not the font-effects tab where color settings are found.

This is related to tdf#111733.

In terms of draw text (like textboxes) and comments
it did nothing at all, so I fixed that too.

I found this looking for other possible uses
for CHAR_DLG_POSITION or CHAR_DLG_EFFECTS.

I didn't find any pre-defined menu where this could
be seen, but a customized menu adding the right
"Font Color" (uno:Color) demonstrates the difference.
There might be more EFFECTS, but this was the only one I found.

Change-Id: If80f14e00d1a2ebae850a885df06fc556f77903a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/129044
Tested-by: Justin Luth <jluth@mail.com>
Reviewed-by: Justin Luth <jluth@mail.com>