ScriptForge - (SF_Platform,SF_FileSystem) Manage extensions
New properties and methods
--------------------------

SF_Platform.Extensions : (property)
   provide a list of the installed extensions
   as an array of strings

SF_FileSystem.ExtensionFolder(Extension as string) : (method)
   provides the location in FileNaming notation of the
   folder where the given extension is installed

The SF_FileSystem.ExtensionsFolder (notice the additional "s")
is kept as a property providing the folder containing the
extensions installed by the user.

All methods and properties are available with an identical
syntax and semantics both from Basic and Python scripts.

The difference between a propert or a method is
irrelevant for Basic but is essential in Python:
a property must not have brackets, a method must have them.
Hence the distinction between
   ExtensionFolder being a method requiring an argument
   ExtensionsFolder being a property without argument

Change-Id: Idd18a5624f4abd84095d146e467e8fdcf497bd0d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/129572
Tested-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jenkins
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>