dynamic logarithmic columns in ScBroadcastAreaSlotMachine
The rows have been handled this way for quite a while, but column
were still using a linear distribution for a hardcoded column limit.
Sheets with >1024 columns didn't work because of the hardcoded limit.

Normal sheets use 57344 slots and 6 distributions (ScSlotData),
and this commit doesn't change that. For huge sheets the original
broken implementation used 90112/10 and this changes to 270336/50
(there are now 5 ScSlotData horizontally instead of just one).
Given that the number of cells is 256x larger I find this acceptable.

Change-Id: I619b979f1363b5427d270f9ca0728415d58f41b4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/129678
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>