simplify autocorrect makefile
The multiple indirections/complex nested constructs don't really provide
any benefit here, they certainly don't make the makefile easier to read.

Also remove the empty mimetype files and just touch them instead of
copying them.

The xml and mimetype files are now intermediates and are removed by make
once the .dat is built. (if that's not desired, they can be flagged as
.SECONDARY)

Further simplification is possible when using the same language code
for both the intermediate directory as well as the source directory
layout and the .dat files

Change-Id: I0e0c28abcad2b404a88d1f32b9b9cdb53b4125c6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/129628
Tested-by: Jenkins
Reviewed-by: Christian Lohmaier <lohmaier+LibreOffice@googlemail.com>