sw a11y: Drop custom implementations of std::less
SwFrameFunc's operator() just applied the less-than operator
which is the same as 'std::less<const SwFrame*>' does.
The latter is the default for the 'Compare' template argument of a
'std::map' with key type 'const SwFrame*>' anyway, so
there's no need to specify any explicitly.

The same applies for 'SwShapeFunc' accordingly.

Change-Id: I2472b38d04dd6d280bfb825d750a6d67c2d9797c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/129682
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>