avoid destructing many unique_ptr's most of which are nullptr
With huge sheets there are so many CellInfo structures that
calling all the destructors actually shows up in profiler data
while trying to scroll around (it allocates CellInfo for every
column starting from 0 up to the last one wanted). Keep
the owning pointers in a vector that'll be much faster to process.

Change-Id: Ibadb05fcce467fb808c38e2ed9a13b4cfef3720e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/129699
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>