sw HTML export: add a new LeadingTabWidth option
This is a simple way to not loose indentation done with tabs (e.g.
source code) during the HTML export.

A more complex way would be ask the layout for the tab portion width,
ask VCL what's the size of an nbsp glyph and then act accordingly, which
is is not done here.

Change-Id: I2a5c0512e9e5541e55e10f29952679bf05d63f1b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/129974
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins